refactor: 요청 uri 수정
