+++ src/main/java/roomescape/controller/ThemeController.java
@@ -2,7 +2,9 @@ package roomescape.controller;
 
 import jakarta.validation.Valid;
 import java.net.URI;
+import java.time.LocalDate;
 import java.util.List;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,7 +12,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.AvailableReservationResponse;
 import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
 import roomescape.service.ThemeService;
@@ -38,6 +42,14 @@ public class ThemeController {
     @GetMapping("/popular")
     public ResponseEntity<List<ThemeResponse>> getTopTenTheme() {
         return ResponseEntity.ok(themeService.getTopTenTheme());
+    }
+
+    @GetMapping("/{id}/times")
+    public ResponseEntity<List<AvailableReservationResponse>> getThemesTimesWithStatus(
+            @PathVariable(name = "id") Long themeId,
+            @RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate date
+    ) {
+        return ResponseEntity.ok(themeService.getThemesTimesWithStatus(themeId, date));
     }
 
     @DeleteMapping("/{id}")
