+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,17 +1,12 @@
 package roomescape.service;
 
-import java.time.LocalDate;
 import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
-import roomescape.dto.AvailableReservationResponse;
 import roomescape.dto.ReservationTimeRequest;
 import roomescape.dto.ReservationTimeResponse;
-import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
 
 @Service
@@ -49,23 +44,6 @@ public class ReservationTimeService {
     public List<ReservationTimeResponse> getReservationTimes() {
         return reservationTimeDao.findAll().stream()
                 .map(ReservationTimeResponse::fromEntity)
-                .toList();
-    }
-
-    public List<AvailableReservationResponse> getReservationTimesWithStatus(Long themeId, LocalDate date) {
-        List<ReservationTime> allTimes = reservationTimeDao.findAll();
-        List<Reservation> reservations = reservationDao.findByThemeIdAndDate(themeId, date);
-
-        Set<Long> bookedTimeIds = reservations.stream()
-                .map(reservation -> reservation.getReservationTime().getId())
-                .collect(Collectors.toSet());
-
-        return allTimes.stream()
-                .map(time -> new AvailableReservationResponse(
-                        time.getId(),
-                        time.getStartAt(),
-                        bookedTimeIds.contains(time.getId())
-                ))
                 .toList();
     }
 }
