+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,21 +3,32 @@ package roomescape.service;
 import java.time.Clock;
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import roomescape.dao.ReservationDao;
+import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
+import roomescape.dto.AvailableReservationResponse;
 import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
+import roomescape.model.Reservation;
+import roomescape.model.ReservationTime;
 import roomescape.model.Theme;
 
 @Service
 public class ThemeService {
     private final ReservationDao reservationDao;
+    private final ReservationTimeDao reservationTimeDao;
     private final ThemeDao themeDao;
     private final Clock clock;
 
-    public ThemeService(ReservationDao reservationDao, ThemeDao themeDao, Clock clock) {
+    public ThemeService(ReservationDao reservationDao,
+                        ReservationTimeDao reservationTimeDao,
+                        ThemeDao themeDao,
+                        Clock clock) {
         this.reservationDao = reservationDao;
+        this.reservationTimeDao = reservationTimeDao;
         this.themeDao = themeDao;
         this.clock = clock;
     }
@@ -55,5 +66,22 @@ public class ThemeService {
         if (!isDeleted) {
             throw new IllegalArgumentException("해당하는 ID가 없습니다.");
         }
+    }
+
+    public List<AvailableReservationResponse> getThemesTimesWithStatus(Long themeId, LocalDate date) {
+        List<ReservationTime> allTimes = reservationTimeDao.findAll();
+        List<Reservation> reservations = reservationDao.findByThemeIdAndDate(themeId, date);
+
+        Set<Long> bookedTimeIds = reservations.stream()
+                .map(reservation -> reservation.getReservationTime().getId())
+                .collect(Collectors.toSet());
+
+        return allTimes.stream()
+                .map(time -> new AvailableReservationResponse(
+                        time.getId(),
+                        time.getStartAt(),
+                        bookedTimeIds.contains(time.getId())
+                ))
+                .toList();
     }
 }
