+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.time.Clock;
 import java.time.Instant;
+import java.time.LocalDate;
 import java.time.ZoneId;
 import java.util.List;
 import javax.sql.DataSource;
@@ -20,7 +21,9 @@ import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import roomescape.dao.ReservationDao;
+import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
+import roomescape.dto.AvailableReservationResponse;
 import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
 
@@ -40,7 +43,8 @@ public class ThemeServiceTest {
                 .build();
         JdbcTemplate jdbcTemplate = new JdbcTemplate(dataSource);
         Clock clock = Clock.fixed(Instant.parse("2023-03-26T08:25:24Z"), ZoneId.systemDefault());
-        service = new ThemeService(new ReservationDao(jdbcTemplate), new ThemeDao(jdbcTemplate), clock);
+        service = new ThemeService(new ReservationDao(jdbcTemplate), new ReservationTimeDao(jdbcTemplate),
+                new ThemeDao(jdbcTemplate), clock);
     }
 
     @Test
@@ -117,5 +121,23 @@ public class ThemeServiceTest {
         // when, then
         assertThatThrownBy(() -> service.deleteTheme(id))
                 .isInstanceOf(IllegalStateException.class);
+    }
+
+    @Test
+    void 선택한_날짜와_테마의_시간과_예약_여부_조회하기() {
+        // given
+        LocalDate date = LocalDate.of(2023, 3, 21);
+        Long themeId = 11L;
+
+        // when
+        List<AvailableReservationResponse> responses = service.getThemesTimesWithStatus(themeId,
+                date);
+        List<Boolean> list = responses.stream()
+                .map(AvailableReservationResponse::isBooked).toList();
+
+        // then
+        assertThat(responses).hasSize(4);
+        assertThat(list.stream().filter(b -> b).count()).isEqualTo(1L);
+        assertThat(list.stream().filter(b -> !b).count()).isEqualTo(3L);
     }
 }
